Adds most CRUD functionality to services resource
Adds views to list existing services, create new ones and update existing
ones. It does not add the ability to delete an existing service as this
is not yet needed.

In doing this a few things have changed.

I have changed the error handlers to return JSON and removed the error
handler html templates. I have done this because at the moment we are
building a JSON API, so our error responses should also be in JSON. I
have also replaced the root view with one that returns a JSON object
that links to the /services resource and removed it's template. We have
an intention to build a hypermedia API, part of that is clients being
able to discover the resources they can interact with on their own.

I have refactored the tests to pull application tests into their own
package and split tests on the /services resource out into their own
module. I've done this to better reflect the structure of the
application in the tests and because the prevous test module was getting
quite large. One strange thing I have done is to use a mixin to provide
some common tests to each endpoint that accepts JSON data. We will see
if this is a model that works well.

I have NOT done any validation of the services JSON itself. We have
another story going on to validate against a JSON schema.