move validation module and add helper method
Move the g6importService module into the lib package and
renamed it to validation. The module should be renamed at the least to
be snake case but whether it lives in the lib package and whether
validation is the right name is up for debate.

Introduce a new function that validates some json against the
schemas and raises a 400 if it fails. This removes the duplication in
the POST and PUT views and follows the pattern that SQLAlchemy-Flask
uses with it's find_or_404 method.