Generate service IDs externally
We have decided to use the IDs generated in the SSP to identify the
services. For the moment we will not have a defined way for this API to
generate IDs itself. As such POSTing doesn't make sense. The import will
use PUT and put the SSP ID in the URL. So, changes include:

- POST /services endpoint now returns a 501 Not Implemented with an
  explanatory message.
- PUT /services/{service_id} can now create new services as well as
  update them.