Add pagination to service listing
After discussion with @TheDoubleK and @minglis we have decided to put
the pagination links in the response under the 'links' section. This is
beacause it's more obvious to a user browsing the API than other
solutions such as Link headers.

Page size is not controllable by the user, it's simpler to just define
it and let them paginate.

Pagination links are only shown if they are valid. For example, no
'prev' pagination link is shown on page one.