Allow certificate to be provided to import script
The staging API uses a self signed SSL certificate which causes requests
to error unless specific actions are taken.

This change allows the user to pass in the certificate that is being
used. The default self signed cert generated by openssl does not have
the subjectAltName set, which requests complains about. However, it does
work.