Add a status endpoint
Add a status endpoint that can be used by the ELB to determine whether
the app is healthy. I have chosen not to touch the database on this
status endpoint as that is shared between all EC2 instances and
therefore a slow database would make all instances look like they're
slow.