Build SQLAlchemy URI from RDS variables
When running a Beanstalk environment the RDS credentials are passed to
the app via a series of environment variables. The SQLAlchemy
environment variable is not provided but it can be created from the RDS
variables. This is a backwards compatible change that will generate the
SQLAlchemy URI from the RDS environment variables if it has not been
explicitly set in the environment.