Add a test for current supplierName write behaviour
JSON schema validation allows any additional fields in PUT requests,
so currently it's possible to add a service with the supplierName field.

This will store the 'supplierName' as well as any other additional
fields to the service data JSON blob.

However the supplierName will be shadowed by the value from the DB
when the service data is requested back from the API.