Remove excessive logging from indexing script
I am adding the re-indexing of G6 products as a Jenkins job, and now we know that the script works it seems unnecessary to output the JSON of every service.

I have left the "processing page" print line in there so that progress can be seen as the job runs.