Change AUTH_TOKENS to DM_AUTH_TOKENS
All the other environment vars we use are prefixed with 'DM_'.  For consistency and ease of checking the current environment for DM apps I think it is worth prefixing all our environment vars with 'DM_'.

Note - this will need changes to be made to the ElasticBeanstalk configuration to reflect the new name.