Adding an update endpoint onto the services API.
URL:
POST /services/{:service_id}

Requires Json of the form:
{
	updater : {
		username : "my username",
		reason: "my reason"
	},
	serviceUpdate: {
		"key": "value"
	}
}

Where serviceUpdate can be any valid JSON including Arrays and Objects etc.

Will load the existing service (404 if not exists) update the JSON data and utilise the JSONParse routines that currently power the import to validate.

If all valid will persist in the database.

This commit contains no DB changes so username and reason are not persisted at this stage.