Preload jsonschema validators
Tests were taking 13s to run, most of the time being spent in
jsonschema. This commit pre-loads the jsonschema validators at startup
rather than just loading the actual JSON at startup. This has two
main benefits. It means that the app will fail to start if the JSON
schemas are invalid and it improves performance considerably. After this
commit the tests take a little over 6s to run and the bottleneck is now
the database.