Updated the services table to have a string type for the service id, G6 services have int IDs, but G5/G4 use strings.
update to this required updates to the validation of the API endpoints that accept service_ids. Service IDs need to be alphanumeric/dashes to be valid.

Updated JSONParse schemas to have strings for IDs.

Changed all the tests to encompass this.

Also added a status field to the services tables, with check constraints to handle publishing/enabling/disabling of services.