Make G6 regexps stricter
Addresses this bug: https://www.pivotaltracker.com/story/show/91384162

Currently regular expressions do not specify start and end of strings, so it would be possible to update a service with values that only partially match the specified strings (e.g. "Months" would be accepted as a value for for 'priceInterval' because it contains "Month", even though it was not an option in the SSP).

By specifying the start and end of string anchors here it will ensure only exact matches to the allowed values will validate.