Allow priceMax to be null
Updates to remove a value of priceMax (an optional field in the SSP) are failing because `None is not of type 'number'`.
This update to the G6 JSON schemas allows null values as well as the number type for priceMax.

We don't need similar changes for other optional fields because they are strings or lists, and can happily be set to an empty string or empty list if an update is made to remove an answer to one of these optional questions.