Remove 'id' field from service JSON data
The 'service_id' database field should be the only place where service_id is stored - it should not also be in the JSON data.
Uploaded services need to specify the ID in the JSON, so that it can be matched to the ID specified in the PUT/POST request (and rejected if they do not match), but the field should be removed before the JSON is stored in the database.