Send application logs to CloudWatch
Also move the apache access logs so that we have better consistend
naming. I tried putting these both in the same file but it bacame a bit
hard to follow.

Creating of the logs directory is done with container commands and
duplication of the path because CloudFormation::Init cannot create just
directories and function calls cannot be used as the file name key.