Increase apache request timeout
We have been seeing a lot of 408 errors in the apache access logs. These
are cause by the ELB healthcheck. Discussion on the internet [1]
suggests this is related to Apache's mod_reqtimeout extension [2].
Rather than completely disabling the request timeout I have increased it
slightly. Testing this shows that it does stop the 408 log messages.

[1]
http://serverfault.com/questions/485063/getting-408-errors-on-our-logs-with-no-request-or-user-agent
[2] http://httpd.apache.org/docs/2.4/mod/mod_reqtimeout.html