Move test db migration to package setup
This speeds the tests up a lot (~45 seconds down to ~16).

Move the running of all migrations and dropping of all tables to a package
setup which means that it's only run once for all tests in that package.
The contents of tables are dropped after each test.

One bad thing in this pull request is that because we bootstrap the
frameworks table in migrations we have to do that bootstrapping manually
in the tests as well.