Make import_service more like update_service
Move code around so that it flows more similarly to update_service.

Dissallow updates via PUT. All our updates currently go through POST and
the import_service method already makes assumptions about it always
creating new services (it always returns a 201 status code and always
updates the created_at field).

Remove some redundant tests that are either in the wrong place (testing
post in the put test) or no longer necesarry (testing the search API is
called on update through the PUT endpoint).

Ensure supplierId does not make it into the data field, it is in the
supplier_id field of the model.