Convert JSON schema patterns to enums
Where possible convert JSON schema string regular expression patterns
into enums.

With the development of the SSP we will need to take a closer look at how we
describe what's needed for a given framework. Currently some of this logic is
duplicated between these schemas and the SSP content repo. Making
'choice' regular expressions enums is intended to make it easier to
reuse this validation logic elsewhere, where providing a choice is
required.