Return 'id' field in ContactInformation.serialize
Returning unique 'id' as part of the contact details allows the client
to address the given contact information object in POST/PUT/GET/DELETE
requests.

Table primary key is used for 'id'.

'id' is also added to the contact-information JSON schema as an optional
field. This means that API-to-API imports will still pass validation
without requiring the API to drop any nested contactInformation fields,
but the ids aren't actually used in PUT endpoint.