Add Supplier and ContactInformation versioning
Uses SQLAlchemy-Continuum to create archived versions of Supplier and
ContactInformation objects on data changes. The versions are created
automatically on INSERT, UPDATE and DELETE queries and are stored in
`<model>_version` table.

Each object can reference previous version using the `.versions`
attribute. Each version contains a `.changeset`, which is a list of
fields that were changed when it was created. Versions can also be
used to replace the current object using `.restore`.

The inital version is created when the object is added for the first
time.

When an object is modified a version record is also created for any
related objects (e.g. modifying supplier will create a new version
for related ContactInformation objects). This shouldn't be an issue
since only direct (not backref) references are traversed, so modifying
supplier won't create archives for all related services.