Store update author when editing supplier details
Adds required 'updated_by' field to the edit request.
Field value is saved to the audit event user field and as part of
the request data.

There's no `update_reason` for the supplier update since it's being
deprecated in other edit endpoints once they're switched to audit
events. `update_reason` was always set to a static value and since
there are no plans to add a reason field to frontend apps it's likely
to stay that way if it's kept.