Move some query logic into Service model
This change fixes a bug in the list services view where the Framework
status filter was not being applied when filtering by status.

This change also moves some of the Service querying logic into the
Service model so that it can be tested in isolation and shared between
multiple places. This logic has been added to the Service rather than
the mixin because the logic is only needed on the Service right now.