Allow a specified list of fields to be required
This modifies the `get_validation_errors` method to allow a list of required fields to be passed in.
This will be useful for validating individual pages of questions - the frontend will be able to pass in the list of fields that were required on a page, and the API can validate that they were submitted with the form.