Store archive id instead of a link in audit event data
Since url_for is using the Host header to build the link, when API
is called from another app the URL is generated using internal DNS
name. Storing links instead of ids would also make it harder to
change the archived-services route.

Instead, only archived-services ids are stored in the audit event
data. When the audit event is serialized to JSON the ids are used
to add [old|new]_archived_service to the list of links.