Fix word-count regular expressions to not hang
The word count regular expression with *s in it hangs when checking text over the allowed word count.
I have tried various possibilities before getting to the ugly regex here.
Using word boundaries (`\b`) doesn't work well because it fails to match strings that begin or
end with a non-alphanumeric character. I have added the `\s?` at the beginning and end of the regex
because it seems wrong to fail to validate items with leading or trailing whitespace.