Refactor SelectionAnswers model
Make the find_by_supplier_and_framework method a static method on the
SelectionAnswers model.

This was done because the query classes are a bit less obvious and
therefore we should avoid them unless we really need them. In this case
we only ever get 0 or 1 instances, while we lose the first_or_404
pattern we avoid a bit of confusion.