Allow features and benefits over 100 characters
While doing the import of services it seems there is at least one with a feature that is over 100 characters long (103 to be precise).
This PR allows the features and benefits fields (for existing G5 and G6 services only) to be up to 120 characters long, but still only up to ten words.

For G7 we will stick to the 100 character limit, as that should be plenty for ten words.