Append ' copy' to the copied draft name
Draft copies are named "<original name> copy" to avoid accidental
duplicate names in submitted services.

Since the name change doesn't go through service validation it's
possible that the new draft name is longer than the serviceName
limit of 100 characters and won't pass validation if the draft
is completed.

To avoid this there's an additional check that adds " copy" only
if the new name will stay below the 100 character limit. Since the
limit can only be increased without breaking existing services this
check should be enough to make sure that the new service name passes
the validation length rule in the future.