Retrieve audit events by object reference
Allow audit events to be retrieved by object reference identified by the
object type and the object ID. The object type is as it appears in the
URL and the JSON payloads as this is on the external interface. If one
of object-type or object-id is provided both must be provided.