Rewrite queries to use all the joins
Services, users and suppliers are the main nouns that we want to know about.
They can have 3, 3 and 2 statuses each.

This commit rewrites the queries to count every combination of these states.
The app can then sum by lot, or status to display the data it actually wants,
without making lots of individual and specific queries.