Add decorator to set PostgreSQL isolation level
Add a Flask decorator to allow the PostgreSQL transaction isolation
level to be set per request.

On some update requests we can run into concurrency issues because we
read a row with a big blob of JSON from the database make some changes
and then write it back. If another process was also doing this, even if
it was in another part of the JSON there would be a race condition and
the last writer would win. PostgreSQL has a transaction isolation level
[1] called SERIALIZABLE that will cause a failure if any of the rows
that a transaction trys to work with are changed while it is running. We
don't want to set this isolation level on all transactions because it's
more costly and often we don't care. This change allows us to set the
isolation level on a single request transaction.

[1] http://www.postgresql.org/docs/9.1/static/transaction-iso.html