Change prices to strings in G5&6 schemas
Change the minPrice and maxPrice fields to be strings (as they are in
G7) in the G5 and G6 schemas.

This is going to mean that G5 and G6 services read from the data API
will not be valid if sent straight back. However, this shouldn't be a
problem because the only way they can be edited is through the admin
app which sends them back as strings.

In terms of the effect this will have on other apps. The supplier and
frontend apps will be using the same price rendering code which will
allow for strings or numbers for the prices. The buyer frontend
currently uses the `priceString` field which isn't set on G7 services so
we need to address this anyway.