Rename selection_answers to supplier_frameworks
Declaration is the name we have settled on for the questions that
a supplier answers to get onto a framework. As there are more details
that we seem to need to hang on this relationship, such as whether
they've registered interest, this has been pulled out into a table
that more generally models the relationship between a supplier and
frameworks.

This change will cause downtime while the code is being deployed. The
code required to make it zero-downtime did not seem worth the effort
when weighed against the fact that this can be done out of hours.