Add `frameworkName`, `frameworkSlug` to fixtures
Tests were failinga ll over the place because they were importing
services without any framework information.
This is wrong, because in our real-world application, services
cannot be created except from a draft service; service imports
are *exclusively* done using services exported from the API.
So it's correct to assume services we're looking to import are
going to have the framework information in them.