Use object relationships to get validation schema name
Instead of making a query for the framework object inside the helper
validation methods use the service object relationships. Framework
and lot can still be passed in as optional arguments for objects
that haven't been committed to the database yet, since they might
not have the related objects (e.g. when creating a new service from
scratch or by copying a draft).