Add service helper to validate framework, lot and supplier fields
When creating a service or draft checks that incoming JSON has a
valid framework and lot slugs and supplier ID and returns the
referenced objects.

Depending on the endpoint logic this could be run before the
JSON schema validation, so the function checks that required
keys exist and handles the non-integers supplier ID DataError.

Even though the supplier object is rarely used by calling views
and we could continue to rely on the IntegrityError to check for
missing suppliers, adding it to the same function simplifies the
logic a bit by creating a single place responsible for verifying
references. The added Supplier query shouldn't add significant
load.

As a side-effect, this changes the error response message for
non-digit supplierId values (used to be JSON schema error) and
missing supplierIds (used to be database IntegrityError).