Remove framework slug from  draft service creation URL
Framework slug should be passed as part of the service data. This
is what the service import endpoint does already and matches the
way we treat lots, supplier IDs and other special fields.

The /draft-services/<framework_slug> URL should be considered deprecated
and will log a warning on each use. Once it's no longer used by the
apiclient it should be removed.