Add update framework endpoint
Since the valid framework statuses are no longer held in the database
and the only way we have of updating the frameworks is to manually
update them in the database this leaves us open to having invalid
statuses. This change introduces an update endpoint that allows the
status to be changed through the model, applying the validation.

It only allows the status to be changed as this is the value that get's
manually changed between environments at the moment.