Don't allow copying a one_service_limit draft
For lots that can only have one service per supplier there shouldn't
be a way to create additional services, so we return an error if the
copy draft endpoint was called.

Since there's no "Make a copy" button displayed for one-service lots
this requires no changes in the frontend apps.