Remove `anyOf` key for not-strict service validations
When validating schemas, we either validate strictly using the whole schema,
or we validate using a subset of the required fields so that we can, for
example, submit only specific fields present on a page.
However, if we don't remove the list of `anyOf` validations, we end up with a
bunch of fields that are unintentionally required.