Return reasonable error message if anyOF error
If our schema fails to validate because a `anyOf` subset is missing,
the original error message is basically useless.
This extra check looks for the `title` key in the first returned
validation error after failing on `anyOf`.
If found, this is what will be returned by the API so our frontend
can make some sense of what went wrong.