Return a reasonable error message if dependencies error
When the schema fails to validate because we've not included all keys from
a group of dependencies, the default message is fairly verbose and not
in line with what our frontend apps would be expecting.
However, we can process 'dependencies' error in a similar way that we process
'required' errors because the implications of either are the same.