Change regex matching to work in python2
If we're missing a dependency, the default response from the API looks like
`'developerLocations' is a dependency of 'developerPriceMin'`
Unfortunately, in python2, this is interpreted as
`u'developerLocations' is a dependency of u'developerPriceMin'`
which looks to me like embedded unicode substrings that were somehow badly
interpreted.
I don't know if it's the right solution, but changing the regex expression to
look for an optional 'u' character works in both Python2 and Python3.