Make stats handle null declarations
There was a bug fixed in the previous commit that would result in JSON
'null' values for the supplier declaration which would in turn cause the
stats queries to fail. As this is a subtle bug and the stats are
depended on they need to be pretty robust. This change excludes JSON
'null' values from the query.