Write and refactor user tests
Testing the user exporting route was pretty involved because
in order to test the various states, we have to insert quite
a bit of information into the database.
Wrote a little suite of utility functions to automate the setup,
and in the process noticed that some of the utility functions
could be used in other user testing classes.
Stopped short of changing any tests that were inserting values
directly into the database, nor did I apply them outside of
the user testing classes because rabbit holes.