Fix audit events object-id and object-type filtering
List audit events endpoint used a join between the audit event
table and the `object-type` model, but didn't check the audit event
`object_type`. This means that eg an event with `object_id=1` could
be joined to Services, Suppliers or Frameworks record depending on
the `object-type` argument, which lead to incorrect audit events
being returned by the endpoint.

Instead of a join query we request the referenced object separately
and then filter the audit events based on that.