Prevent buyer user being created with invalid email domain
[#111536536](https://www.pivotaltracker.com/story/show/111536536)

Buyers must be from allowed public sector organisations. This change
prevents buyer user accounts being created unless their email address is
either on one of the allowed domains or on a subdomain.

Allowed domains are stored in a text file rather than handled with a
JSON schema to make it easier to manage the list.