Add new endpoint to update the status of a draft
Currently we only have an endpoint to mark a service as complete, but that is the only change to `status` that can be made.

We now need to mark drafts as `failed`, so this commit adds a new endpoint hat can be used to update the status of a draft to any valid status (valid values are as listed in `DraftService.STATUSES`).