Add framework and lot to brief
This commit adds framework and lot references to the Brief model as a
brief is for a specific framework lot. In doing this we added a foreign
key constraint to ensure that the lot added to a brief is associated to
the provided framework. As this is also useful for services the same
constraint has been retrofitted to the ServiceTableMixin.

The migrations have been kept as three separate migrations as they're
each doing a distinct change.