Add framework.framework to serialized output
For briefs and services (ie. services, draft_services and
archived_services) add the frameworkFramework field to the output. This
is required for the admin app to be able to generate buyer app service
URLs in a framework agnostic way.

The key frameworkFramework is not nice but it follows the existing
pattern. This pattern should probably be changed to include
sub-resources as sub-objects rather than having one big flat key space.