Use JSONUpdateTestMixin to test for missing update details
Remove ad-hoc endpoint tests in favour of JSONUpdateTestMixin test
for missing update details.

Clean-up some duplicate tests and set endpoint and method on update
view tests. Add support for using object attributes in endpoint, to
make it possible to use the mixin with test classes that create the
relevant object during test setup.

This updates most of the view tests with the exception of drafts and
briefs. Briefs and drafts don't split tests between classes for each
endpoint, so setting JSONUpdateTestMixin with one endpoint and method
won't have the same benefit.